+++ clients/journey-client/src/openapi.d.ts
@@ -363,6 +363,10 @@ declare namespace Components {
              * }
              */
             ValidationRuleRef;
+            /**
+             * Manifest/Blueprint ID used to create/update the entity
+             */
+            _manifest?: string /* uuid */[];
             createdBy?: string;
             updatedBy?: string | null;
             /**
@@ -611,6 +615,10 @@ declare namespace Components {
              * }
              */
             ValidationRuleRef;
+            /**
+             * Manifest/Blueprint ID used to create/update the entity
+             */
+            _manifest?: string /* uuid */[];
             createdBy?: string;
             updatedBy?: string | null;
             /**