+++ packages/epilot-sdk-v2/src/types/data-governance.d.ts
@@ -1,4 +1,4 @@
-/* Auto-copied from data-management-client */
+/* Auto-copied from data-governance-client */
 import type {
   OpenAPIClient,
   Parameters,