+++ clients/data-governance-client/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## 0.1.0
-- Initial scaffold of `@epilot/data-management-client`
+- Initial scaffold of `@epilot/data-governance-client`
 - Added OpenAPI generation scripts (including local path)
 - Added build/test configuration and client stubs